Improve clarity of installation instructions
Hello @stefankroes,

I was installing `ancestry` today and found the installation instructions to be a little unclear since they did not differentiate very well between text that should be run at the command line and text that should be added to a file. I have attempted to make the instructions more clear by using code blocks and a `$` to signify commands to be run at the command line and using codeblocks with surrounding context to indicate what files should look like.

Let me know if you think this improves clarity.