Improve performance of arrange_nodes.
The old code is essentially an insertion algorithm that walks the tree
from the root for each node and inserts itself in the right spot.
Inserting a node is roughly O(depth); best case O(1), but worst case
O(n).  Thus, the tree building is roughly O(n * depth); best case O(n),
but worst case O(n^2).

The new code maintains an index of a node's children by node id.  Each
nodes is placed into the children of the parent id. Thus, inserting a
node is an O(1) lookup, making the tree building O(n).

The operation only has to look at the parent, and does not have to go
through all of the ancestry ids, because eventually the parent node
will go through the same insertion process itself.

Note that the old algorithm used a pre-sorting of nodes in order to
more efficiently arrange them.  The new algorithm doesn't care about
order, and thus that sorting can probably be removed in future PRs.