Update README.md
### What is this PR?

* Add detail re: customizing column names.

### Why is it required?

Because it may not be convenient to name the column: 'ancestry' - I imagine (perhaps incorrectly) that many would prefer to customize the column name to something better suited to their use-cases (e.g. https://github.com/stefankroes/ancestry/issues/273).

I hope this change is helpful - if it is not, feel free to close without compunction etc.