set ancestry column collation for tests
Currently, we are using the default database collation for the ancestry column.
In most rails apps, this is some form of unicode locale.

This is skipping indexes for like.
On ubuntu systems, it is ignoring slashes for sorting, which causes all sorts of problems

This changes tests to just use a simple binary comparison - ascii byte for byte.
This gets better results, uses indexes for LIKE, and is faster for all comparisons.

ENV["ANCESTRY_COLLATION"] should never be needed unless you are testing performance characteristics