extract flatten_arranged_nodes
this is the part responsible for taking a tree and producing an array of nodes
with the parent node before the children

Since children is always a hash, this will never enter the sorting code
and never require the block. so the sorting and the &block was dropped.

This works because ruby enumerates the hashes with insert order (no need to sort again)