Drop ancestry_base_class
active_record defines base_class as the top of an STI tree.

ancestry had a similar concept. It was the class where ancestry was defined.
Since ancestry originally called it a base_class,
there was a conflict and in 4e57f32e1758 it was renamed to ancestry_base_class

PRs of note:
- https://github.com/stefankroes/ancestry/pull/86
- https://github.com/stefankroes/ancestry/issues/87

The ancestry_base_class is used quite often to make sure the proper class
is accessed to lookup values.

I'm not sure the use case for defining ancestry in the middle of a hierarchy,
but it breaks a few things.

If you define ancestry in the middle of a tree, the parent classes still
implement all of ancestry.

If these parent classes are added to the tree then the cache counters
and finders will not work properly.

I'm dropping this custom concept and using the standard rails concept
instead, as I think it should have been done originally.