Changes to update_descendants_with_new_ancestry
- this is called after a save (/via #589 ), dropping `new_record?`
- putting `ancestry_changed?` in the callback definition. not needed in the method
- had wanted to remove `sane_ancestor_ids?` since validations already ran
  Unfortunately, `update_attribute` (read: no validations) allows bogus ancestry through.

Decided to continue skipping updates if the ancestry is deemed not valid.
It can only get this way if a user skips validations.

Not sure the desired behavior if something is corrupting ancestor's ancestry value.
I can see reason why we may want to update them all.
For now, keeping this skip updates check in place