test in_database and children throughout the lifecycle
assert_attributes really changed
there are so many attributes to test for each node. It made the tests unreadable.
This required changing the format of the parameters

The matrix felt like a hack at first, but it is not programmatically testing stuff.
there is a hardcoded list of what needs to be done for each node. Which in essence
is what tests do.

This is now properly testing all these attributes for in_database along the whole lifecycle
of these objects