fix touch tests
some tests were not reloading the records to pickup changes
some tests were starting with a recent updated_at. so they always showed as changed

in tests, fixed strange bug in mysql and FORMAT=materialized_path only:
if an update_at is recent, mysql will not send the value to the database to change the updated_at
In our test case, the local updated_at is now, and the database has an updated_at as a while ago
so it is not sending the new update_at to the db.

It looks like rails 7.1.0.a has this fixed (not sure when)
But changed update statements to include a reload so rails knows to send this column across.

This feels like a test only issue so I'm ok modifying only the tests.