Implement depth as a subquery
The main reason I introduced depth sql is to improve rebuild_depth_cache_sql!
Throwing this into the scopes is a bonus. (but not really)

This works great for the update, but the extra scopes will not perform well.
If you need to use them, consider adding an index on that equation