configure ancestry_column name from matrix
Things changed a bit here.
In tests, we are now always passing ancestry_column into has_ancestry (in with_model)
This value is from an env variable or defaulting to ancestry

So when code calls has_ancestry, we're not sure the name of the field in the database
so we need to be more careful.

Mostly noticeable in the tests that manually call has_ancestry
And the test of the default ancestry column name (which assumed has_ancestry was being called without values)